requirements.txt
python-telegram-bot==13.0.0
pymongo==3.11.3
python-dotenv==0.15.0
python-dateutil==2.8.1
requests==2.24.0
gunicorn==20.0.4
dnspython==1.16.0
uuid==1.30
py7zr==0.10.1
pyotp==2.6.0
google-api-core==2.0.0
google-api-python-client==2.18.0
google-auth==2.0.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0